engine: move loop-dominance types out of layout into ltm_dominance
The dominant-period machinery -- FeedbackLoop, DominantPeriod, the
3-way display LoopPolarity, PartitionSurface, and
calculate_dominant_periods -- lived in layout::metadata for historical
reasons (it was built for the diagram pipeline first), which left the
LTM/analysis surface importing its public result types from a layout
submodule: analysis::analyze_model, the FFI discovery result, and
simlin-mcp-core all reached through crate::layout::metadata. Layout
consuming LTM analysis is the right dependency direction; the reverse
was not.

The types and the selection function now live in src/ltm_dominance.rs,
a sibling of ltm_post/ltm_finding, with their tests. layout::metadata
keeps only the layout-facing bundle (ComputedMetadata, StockFlowChain)
and imports the moved types like any other consumer; there are no
compatibility re-exports, so the module has exactly one public path.
No behavior change -- the move is textual, pinned by the unchanged
test suites.
